requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
click==8.1.3
cloudinary==1.30.0
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.2.6
Django==4.1.2
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.88.0
Flask==2.2.2
gunicorn==20.1.0
h11==0.14.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
starlette==0.22.0
tqdm==4.62.3
typing_extensions==4.4.0
urllib3==1.26.8
uvicorn==0.20.0
Werkzeug==2.2.2
whitenoise==6.3.0